Big Initial Commit
StageA: Sorry, forgot that we were submitting via a github link, so my initial progress is rather congested.  This initial commit includes an asset store pack for the main idle character with implementations of the stat systems as well as their effect on HP. Currently, there are still tweaks that need to be made to the math and being able to undo or alter certain stats.